clip-captioner/requirements.txt
torch>=1.10
pytorch3d
open_clip_torch
tqdm
numpy
Pillow
scikit-learn
matplotlib 